fix: moves controller, model into domain
